Fix numeric decode crash in retailers price endpoint
current_price in /api/prices/:id/retailers was returned as a string by
the pg driver (numeric type), causing Swift JSON decoding to fail.
Added ::float cast to match the fix already applied in products.js.

+++ src/routes/prices.js
@@ -47,7 +47,7 @@ router.get('/:productId/history', authenticate, async (req, res) => {
 router.get('/:productId/retailers', authenticate, async (req, res) => {
   try {
     const result = await query(
-      `SELECT id, retailer_name, product_url, current_price, currency, in_stock, last_checked
+      `SELECT id, retailer_name, product_url, current_price::float, currency, in_stock, last_checked
        FROM product_retailers
        WHERE product_id = $1
        ORDER BY current_price ASC NULLS LAST`,
